+++ src/text.h
@@ -68,7 +68,7 @@ struct kmscon_text_ops {
 	void (*unset) (struct kmscon_text *txt);
 	int (*prepare) (struct kmscon_text *txt);
 	int (*draw) (struct kmscon_text *txt,
-		     uint32_t id, const uint32_t *ch, size_t len,
+		     uint64_t id, const uint32_t *ch, size_t len,
 		     unsigned int width,
 		     unsigned int posx, unsigned int posy,
 		     const struct tsm_screen_attr *attr);
@@ -93,7 +93,7 @@ unsigned int kmscon_text_get_rows(struct kmscon_text *txt);
 
 int kmscon_text_prepare(struct kmscon_text *txt);
 int kmscon_text_draw(struct kmscon_text *txt,
-		     uint32_t id, const uint32_t *ch, size_t len,
+		     uint64_t id, const uint32_t *ch, size_t len,
 		     unsigned int width,
 		     unsigned int posx, unsigned int posy,
 		     const struct tsm_screen_attr *attr);
@@ -101,7 +101,7 @@ int kmscon_text_render(struct kmscon_text *txt);
 void kmscon_text_abort(struct kmscon_text *txt);
 
 int kmscon_text_draw_cb(struct tsm_screen *con,
-			uint32_t id, const uint32_t *ch, size_t len,
+			uint64_t id, const uint32_t *ch, size_t len,
 			unsigned int width,
 			unsigned int posx, unsigned int posy,
 			const struct tsm_screen_attr *attr,
