+++ src/text_bblit.c
@@ -62,7 +62,7 @@ static int bblit_set(struct kmscon_text *txt)
 }
 
 static int bblit_draw(struct kmscon_text *txt,
-		      uint32_t id, const uint32_t *ch, size_t len,
+		      uint64_t id, const uint32_t *ch, size_t len,
 		      unsigned int width,
 		      unsigned int posx, unsigned int posy,
 		      const struct tsm_screen_attr *attr)
